J. Nepomucenus: psal1742 processed
this one is interesting, as it's quite old,
has some relation to the Czech lands (although it probably wasn't all
too widely used, given the amount of exemplars in libraries)
and the tunes are quite singable.
Also the multiple identified similarities to antiphons in various later
publications are worth attention.